requirements.txt
dagster
dagster-webserver
cloudscraper
beautifulsoup4
PyMuPDF
python-docx
python-pptx
chromadb
langchain
langchain-chroma
langchain-groq
langchain-ollama
ollama
pandas
python-dotenv
gradio
groq
matplotlib
langgraph
